+++ CHANGELOG.md
@@ -4,20 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
 ### Added
 - Provide routed RPC capabilities [\#26](https://github.com/angiolep/akka-wamp/issues/26)
   - Client API now provides ``register`` and ``invoke`` operations
   - Router now able to register procedures and route invocations to clients
+- Validate dictionaries [\#19](https://github.com/angiolep/akka-wamp/issues/19)      
+
 ### Changed
 - Improve ScalaDoc comments [\#23](https://github.com/angiolep/akka-wamp/issues/23)
+- Improve ReadTheDocs
+
 
 ## [0.7.0] (2016-09-12)
+
 ### Added
 - Provide better API for payloads [\#16](https://github.com/angiolep/akka-wamp/issues/16)
 - Parse incoming messages skipping the payload [\#17](https://github.com/angiolep/akka-wamp/issues/17)
 
 
 ## [0.6.0] (2016-08-24)
+
 ### Added
 - Validate against loose URIs by default but strict URIs validation now configurable [\#7](https://github.com/angiolep/akka-wamp/issues/7)
 - Auto-create realms by default but abort session (when client requests unknown realm) now configurable
@@ -27,11 +34,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 
 ## [0.5.1] (2016-08-21)
+
 ### Added
 - Provide proper error handling in ``router.Transport`` [\#18](https://github.com/angiolep/akka-wamp/issues/18)
 
 
 ## 0.5.0 (2016-08-20)
+
 ### Added
 - WAMP Router with limited features
 - Future based API for writing WAMP Clients with limited features